pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-simple-pubmed"
version = "0.1.5"
description = "An MCP server that provides access to PubMed articles through Entrez API."
readme = "README.md"
authors = [
    {name = "Andy Brandt", email = "andy@codesprinters.com"},
    {name = "Claude", email = "claude@anthropic.com"}
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp",
    "biopython",      # for PubMed/Entrez API access
    "trafilatura",    # for extracting text from web pages
    "httpx",          # for making HTTP requests with async support
]

[project.urls]
Homepage = "https://github.com/andybrandt/mcp-simple-pubmed"

[project.scripts]
mcp-simple-pubmed = "mcp_simple_pubmed:main"